restructure thrust code but move old thrust code to old folder
